[IN-63] Disable MINTER account (backoffice) (#53)
* feat: add isActivated column on User table

* feat: disable minter account

* feat: display error if account is disabled on login

* refactor: create hook to centralize minter actions

* feat: translate user not activated message

* feat(minter): protect delete endpoint

* chore: apply PR review requested changes

+++ src/app/api/minters/[id]/route.ts
@@ -1,6 +1,9 @@
 import deleteMinter from "@/lib/query/minters/deleteMinter"
 import getMinterById from "@/lib/query/minters/getMinterById"
-import { StatusCodes } from "http-status-codes"
+import isAdmin from "@/lib/utils/auth/isAdmin"
+import isAuthenticated from "@/lib/utils/auth/isAuthenticated"
+import isCurrentUser from "@/lib/utils/auth/isCurrentUser"
+import { ReasonPhrases, StatusCodes } from "http-status-codes"
 
 export async function GET(req: Request, { params }: { params: { id: string } }) {
   const { id } = params
@@ -10,7 +13,21 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
 }
 
 export async function DELETE(req: Request, { params }: { params: { id: string } }) {
+  const session = await isAuthenticated()
+
+  if (!session) {
+    return Response.json(
+      { message: ReasonPhrases.UNAUTHORIZED },
+      { status: StatusCodes.UNAUTHORIZED }
+    )
+  }
+
   const { id } = params
+
+  if (!isCurrentUser(id, session) && !isAdmin(session)) {
+    return Response.json({ message: ReasonPhrases.FORBIDDEN }, { status: StatusCodes.FORBIDDEN })
+  }
+
   await deleteMinter(id)
 
   return new Response(null, { status: StatusCodes.NO_CONTENT })
